#1 Implemented Interface logic for Grid
new class hierarchy:
Interface IGrid is implemented by
GridMap:ScriptableObject which is called by
GridController:Monobehaviour

advantage of this method is that the monobehavior class does not have wrapper methods, so the code looks cleaner. Also, Allows the game to be moddable. Creating assets becomes easier as well and can allow the game dev to create different sized grids.