#1 Refactored hexTile interface
new calss hierarchy:

IHexTile is an interface implemented by hexTile.cs, which contains a hexTileDefinition, which is any shared asset amongst all tiles. this means that materials that are common to all of them need to be instantiated only once. also, code is more readable by giving each material its own name. HexTileManager is the monobehaviour controller.
Moved the feature to only click one tile at a time to the tile class, in order to keep seperation of concerns. makes more sense to have the tile calss take care of tile logic.